設定を更新

+++ packages/user_repository/lib/src/entities/user_entity.dart
@@ -34,7 +34,7 @@ class MyUserEntity {
         userId: doc['userId'],
         email: doc['email'],
         name: doc['name'],
-        age: doc['age'],
+        age: doc['age'] as int,
         description: doc['description'],
         location: doc['location'],
         pictures: doc['pictures']);
